+++ feedme/templates/feedme/react/index.html
@@ -49,7 +49,6 @@ Feedme in React.js
         }
     });
     </script>
-    <script type="text/babel" src="{{ STATIC_URL }}js/feedme.js"></script>
 {% endblock %}
 
 {% block submenu %}
@@ -60,4 +59,5 @@ Feedme in React.js
 
     <div id="feedme-main"></div>
 
+    <script src="{{ STATIC_URL }}js/feedme.min.js"></script>
 {% endblock %}
